feat: windowed ratio decline for confabulation detection + lint/mypy fixes
Replace _consecutive_ratio_declines with _windowed_ratio_declines for
confabulation trajectory detection. Counts total declining steps in a
sliding window (default 5) instead of requiring strict consecutive
declines. This catches V-shaped recovery patterns (Gemini), late-onset
confabulation, and threshold-boundary bouncing.

New config: source_finding_ratio_decline_window (default 5, env
VITALS_SOURCE_FINDING_RATIO_DECLINE_WINDOW). Plumbed through all config
paths (yaml, env, dict, profile).

Bench validation: window=5 yields +7 TP (222→229), P_lb=0.828,
R_lb=0.794 — all gates pass.

Added 12 tests for _windowed_ratio_declines (11 unit + 1 integration).
Fixed 4 unused imports in test_signal_mapping.py and 1 mypy return type
in signal_mapping.py.

+++ agent_vitals/detection/signal_mapping.py
@@ -134,7 +134,7 @@ def _coefficient_of_variation(values: Sequence[float]) -> float:
     if avg <= 0.0:
         return 0.0
     variance = sum((v - avg) ** 2 for v in values) / len(values)
-    return (variance**0.5) / avg
+    return float((variance**0.5) / avg)
 
 
 __all__ = [
